# app/static/swagger.yaml
swagger: "2.0"
info:
  title: "MechanicShop API"
  description: "This is an API to manage a Mechanic Shop"
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
paths:
  /customers/login:
    post:
      tags:
        - Customers
      summary: "Login endpoint (generates token)"
      description: "Validates customer credentials, generates and returns a customer-specific token."
      parameters:
        - in: "body"
          name: "body"
          description: "Login Credentials"
          required: true
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: "Login successful"
          schema:
            $ref: "#/definitions/LoginResponse"
  /customers:
    post:
      tags:
        - Customers
      summary: "Create Customer"
      description: "Create a new customer."
      parameters:
        - in: "body"
          name: "body"
          description: "Customer details"
          required: true
          schema:
            $ref: "#/definitions/CreateCustomerPayload"
      responses:
        201:
          description: "Customer created successfully"
          schema:
            $ref: "#/definitions/Customer"
    get:
      tags:
        - Customers
      summary: "Get Customers"
      description: "Retrieve a list of customers with optional pagination."
      parameters:
        - name: "page"
          in: "query"
          type: "integer"
          description: "Page number"
        - name: "per_page"
          in: "query"
          type: "integer"
          description: "Number of customers per page"
      responses:
        200:
          description: "List of customers"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Customer"

    put:
      tags:
        - Customers
      summary: "Update Customer"
      description: "Updates the authenticated customer's information. Requires token authentication."
      security:
        - bearerAuth: [] # Assuming token-based authentication
      parameters:
        - in: "body"
          name: "body"
          description: "Customer details to update"
          required: true
          schema:
            type: object
            properties:
              name:
                type: string
                description: "The customer's name."
                example: "John Doe"
              email:
                type: string
                description: "The customer's email address."
                example: "john.doe@example.com"
              phone:
                type: string
                description: "The customer's phone number."
                example: "+1234567890"
              password:
                type: string
                description: "The customer's password."
                example: "securepassword123"
            required:
              - email
              - phone
      responses:
        200:
          description: "Customer updated successfully."
          schema:
            type: object
            properties:
              id:
                type: integer
                description: "The customer's ID."
                example: 1
              name:
                type: string
                description: "The customer's name."
                example: "John Doe"
              email:
                type: string
                description: "The customer's email address."
                example: "john.doe@example.com"
              phone:
                type: string
                description: "The customer's phone number."
                example: "+1234567890"
        400:
          description: "Validation error or email/phone conflict."
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Email or phone number already in use."
        404:
          description: "Customer not found."
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Invalid customer ID."
        401:
          description: "Unauthorized access."
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Unauthorized."

    delete:
      tags:
        - Customers
      summary: "Endpoint to delete a customer"
      description: "Deletes a customer. Requires token authentication, and the customer must be logged in to delete their account."
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Successfully deleted the customer"
          schema:
            $ref: "#/definitions/DeleteCustomerResponse"
          examples:
            application/json:
              message: "Successfully deleted customer <customer_id>."
        404:
          description: "Invalid customer ID"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Invalid customer ID"
        401:
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Unauthorized"
  /customers/{customer_id}:
    get:
      tags:
        - Customers
      summary: "Get Customer by ID"
      description: "Retrieve a customer by their ID."
      parameters:
        - name: "customer_id"
          in: "path"
          required: true
          type: "integer"
          description: "ID of the customer"
      responses:
        200:
          description: "Customer details"
          schema:
            $ref: "#/definitions/Customer"
        404:
          description: "Customer not found"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Customer not found"
...
  /inventory/items:
    post:
      tags:
        - Inventory
      summary: "Add to Inventory"
      description: "Add a new item to the inventory."
      parameters:
        - in: "body"
          name: "body"
          description: "Inventory item details"
          required: true
          schema:
            $ref: "#/definitions/InventoryItemPayload"
      responses:
        201:
          description: "Item added to inventory successfully"
          schema:
            $ref: "#/definitions/InventoryItem"
    get:
      tags:
        - Inventory
      summary: "Get Parts with Pagination"
      description: "Retrieve inventory items with optional filters and pagination."
      parameters:
        - name: "min_price"
          in: "query"
          type: "number"
          description: "Minimum price filter"
        - name: "max_price"
          in: "query"
          type: "number"
          description: "Maximum price filter"
        - name: "unassignedParts"
          in: "query"
          type: "boolean"
          description: "Filter for unassigned parts"
        - name: "page"
          in: "query"
          type: "integer"
          description: "Page number"
        - name: "per_page"
          in: "query"
          type: "integer"
          description: "Number of items per page"
      responses:
        200:
          description: "List of inventory items"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/InventoryItem"
  /inventory/items/{item_id}:
    delete:
      tags:
        - Inventory
      summary: "Remove Inventory"
      description: "Remove an item from the inventory by its ID."
      parameters:
        - name: "item_id"
          in: "path"
          required: true
          type: "integer"
          description: "ID of the inventory item"
      responses:
        200:
          description: "Item removed from inventory successfully"
    put:
      tags:
        - Inventory
      summary: "Update Part"
      description: "Update the details of an inventory item."
      parameters:
        - name: "item_id"
          in: "path"
          required: true
          type: "integer"
          description: "ID of the inventory item"
        - in: "body"
          name: "body"
          description: "Updated inventory item details"
          required: true
          schema:
            $ref: "#/definitions/InventoryItemPayload"
      responses:
        200:
          description: "Inventory item updated successfully"
          schema:
            $ref: "#/definitions/InventoryItem"
  /inventory/assign-item/{item_id}:
    post:
      tags:
        - Inventory
      summary: "Add Parts To Service"
      description: "Assign parts from inventory to a service ticket."
      parameters:
        - name: "item_id"
          in: "path"
          required: true
          type: "integer"
          description: "ID of the inventory item"
        - in: "body"
          name: "body"
          description: "Details of the assignment"
          required: true
          schema:
            $ref: "#/definitions/AssignItemPayload"
      responses:
        200:
          description: "Parts assigned to service ticket successfully"
...
definitions:
  LoginCredentials:
    type: "object"
    properties:
      email:
        type: "string"
      password:
        type: "string"
    required:
      - email
      - password
    example:
      email: "Tom@Tom.com"
      password: "1234"
  LoginResponse:
    type: "object"
    properties:
      token:
        type: "string"
      message:
        type: "string"
      status:
        type: "string"
    example:
      token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      message: "Login successful"
      status: "success"
  Customer:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
    example:
      id: 1
      name: "John Doe"
      email: "johndoe@example.com"
      phone: "123-456-7890"
  CreateCustomerPayload:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"
    required:
      - name
      - email
      - phone
      - password
    example:
      name: "Tom Tom"
      email: "Tom@Tom.com"
      phone: "987-654-3210"
      password: "1234"
...
  InventoryItem:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      price:
        type: "number"
    example:
      id: 10
      name: "Brake Pads"
      price: 75.50
  InventoryItemPayload:
    type: "object"
    properties:
      name:
        type: "string"
      price:
        type: "number"
    required:
      - name
      - price
    example:
      name: "Air Filter"
      price: 25.00
  AssignItemPayload:
    type: "object"
    properties:
      service_ticket_id:
        type: "integer"
      quantity:
        type: "integer"
    required:
      - service_ticket_id
      - quantity
    example:
      service_ticket_id: 101
      quantity: 2
securityDefinitions:
  bearerAuth:
    type: "apiKey"
    name: "Authorization"
    in: "header"
...